requirements.txt
# HealthCare-AI-Quantum-System Dependencies
# Python 3.11+

# ============================================================================
# CORE FRAMEWORKS
# ============================================================================
fastapi==0.109.0
uvicorn[standard]==0.27.0
pydantic==2.5.3
pydantic-settings==2.1.0

# ============================================================================
# AI/ML - LLMs
# ============================================================================
openai==1.10.0
anthropic==0.8.1
google-generativeai==0.3.2
langchain==0.1.4
langchain-openai==0.0.5
langchain-anthropic==0.0.1

# ============================================================================
# AI/ML - MEDICAL MODELS
# ============================================================================
transformers==4.36.2
torch==2.1.2
torchvision==0.16.2
monai==1.3.0
scikit-learn==1.4.0
xgboost==2.0.3
lightgbm==4.3.0

# ============================================================================
# COMPUTER VISION
# ============================================================================
opencv-python==4.9.0.80
pillow==10.2.0
pydicom==2.4.4
SimpleITK==2.3.1

# ============================================================================
# QUANTUM COMPUTING
# ============================================================================
qiskit==0.45.3
qiskit-ibm-runtime==0.17.0
qiskit-aer==0.13.2
qiskit-optimization==0.6.0

# ============================================================================
# DATABASES
# ============================================================================
asyncpg==0.29.0
psycopg2-binary==2.9.9
sqlalchemy==2.0.25
alembic==1.13.1
motor==3.3.2  # Async MongoDB
pymongo==4.6.1
redis==5.0.1

# ============================================================================
# DATA PROCESSING
# ============================================================================
pandas==2.1.4
numpy==1.26.3
scipy==1.11.4
polars==0.20.3  # Fast dataframes

# ============================================================================
# MESSAGE QUEUE & STREAMING
# ============================================================================
kafka-python==2.0.2
aiokafka==0.10.0
pika==1.3.2  # RabbitMQ

# ============================================================================
# SECURITY & COMPLIANCE
# ============================================================================
cryptography==42.0.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
pyjwt==2.8.0

# ============================================================================
# HEALTHCARE STANDARDS
# ============================================================================
fhir.resources==7.1.0
hl7==0.4.5
hl7apy==1.3.4

# ============================================================================
# MONITORING & OBSERVABILITY
# ============================================================================
prometheus-client==0.19.0
opentelemetry-api==1.22.0
opentelemetry-sdk==1.22.0
sentry-sdk==1.40.0

# ============================================================================
# UTILITIES
# ============================================================================
python-dotenv==1.0.0
pyyaml==6.0.1
httpx==0.26.0
aiohttp==3.9.1
tenacity==8.2.3  # Retry logic
loguru==0.7.2
arrow==1.3.0  # Better datetime
python-dateutil==2.8.2

# ============================================================================
# VALIDATION & TESTING
# ============================================================================
pytest==7.4.4
pytest-asyncio==0.23.3
pytest-cov==4.1.0
faker==22.2.0
hypothesis==6.98.1

# ============================================================================
# CODE QUALITY
# ============================================================================
black==24.1.1
ruff==0.1.14
mypy==1.8.0
pre-commit==3.6.0

# ============================================================================
# DEPLOYMENT
# ============================================================================
gunicorn==21.2.0
docker==7.0.0
kubernetes==29.0.0
